refactor(hwp5): extract text/shapes clusters from projection.rs (E7 #3)
projection.rs(4,719줄)에서 얽힘이 적은 leaf 클러스터 2개만 보수적 추출:
- projection/text.rs: 텍스트/run 분할(split_text_by_runs, utf16_boundaries,
  char_shape_* 등 7 fn + 단위 테스트)
- projection/shapes.rs: GSO 도형 투영(project_*_run + Axis + scale_* + 테스트)
코어 상태머신(3 진입점 + 공유 상태 + image/field/table/paragraph)은
mod.rs 유지. 이동 fn 은 private->pub(super) (crate-내부, 외부 API 불변).

순수 기계적 분할 — 동작·공개 API byte-동일. make ci 2536 통과.
